+++ instrumentation/opentelemetry-instrumentation-logging/src/opentelemetry/instrumentation/logging/__init__.py
@@ -15,6 +15,7 @@
 # pylint: disable=empty-docstring,no-value-for-parameter,no-member,no-name-in-module
 
 import logging  # pylint: disable=import-self
+from itertools import chain
 from os import environ
 from typing import Collection
 
@@ -29,6 +30,10 @@ from opentelemetry.instrumentation.logging.environment_variables import (
     OTEL_PYTHON_LOG_LEVEL,
 )
 from opentelemetry.instrumentation.logging.package import _instruments
+from opentelemetry.sdk._logs import LoggingHandler
+from opentelemetry.sdk.environment_variables import (
+    _OTEL_PYTHON_LOGGING_AUTO_INSTRUMENTATION_ENABLED,
+)
 from opentelemetry.trace import (
     INVALID_SPAN,
     INVALID_SPAN_CONTEXT,
@@ -77,6 +82,7 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
 
     _old_factory = None
     _log_hook = None
+    _instrumented_loggers = None
 
     def instrumentation_dependencies(self) -> Collection[str]:
         return _instruments
@@ -126,6 +132,7 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
             return record
 
         logging.setLogRecordFactory(record_factory)
+        self._instrument_loggers(**kwargs)
 
         set_logging_format = kwargs.get(
             "set_logging_format",
@@ -146,7 +153,98 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
 
             logging.basicConfig(format=log_format, level=log_level)
 
+    @staticmethod
+    def _instrument_loggers(**kwargs):
+        def get_propagate(logger) -> bool:
+            return logger._otel_propagate
+
+        def set_propagate(logger, value):
+            old_value = True
+            try:
+                old_value = logger._otel_propagate
+            except AttributeError:
+                pass
+
+            if old_value != value:
+                if value:
+                    old_otel_handler = None
+                    try:
+                        old_otel_handler = logger._otel_handler
+                    except AttributeError:
+                        pass
+                    if old_otel_handler is not None:
+                        logger.removeHandler(old_otel_handler)
+                    logger._otel_handler = None
+                else:
+                    # If the handler is already there, skip instrumentation
+                    if any(
+                        (
+                            isinstance(handler, LoggingHandler)
+                            for handler in logger.handlers
+                        )
+                    ):
+                        logger._otel_handler = None
+                    else:
+                        logger._otel_handler = LoggingHandler()
+                        logger.addHandler(logger._otel_handler)
+            logger._otel_propagate = value
+
+        logging_enabled = kwargs.get(
+            "logging_enabled",
+            environ.get(
+                _OTEL_PYTHON_LOGGING_AUTO_INSTRUMENTATION_ENABLED, "false"
+            ).lower()
+            == "true",
+        )
+
+        if logging_enabled:
+            for instrument_logger in chain(
+                (logging.getLogger(),),
+                (
+                    logging.getLogger(name)
+                    for name in logging.root.manager.loggerDict
+                ),
+            ):
+                if not isinstance(instrument_logger, logging.Logger):
+                    continue
+                # Set the _otel_propagate attribute and install handler if necessary
+                set_propagate(instrument_logger, instrument_logger.propagate)
+                del instrument_logger.propagate
+
+            logging.Logger.propagate = property(get_propagate, set_propagate)
+            LoggingInstrumentor._instrumented_loggers = True
+
     def _uninstrument(self, **kwargs):
+        self._uninstrument_loggers()
         if LoggingInstrumentor._old_factory:
             logging.setLogRecordFactory(LoggingInstrumentor._old_factory)
             LoggingInstrumentor._old_factory = None
+
+    @staticmethod
+    def _uninstrument_loggers():
+        if LoggingInstrumentor._instrumented_loggers:
+            LoggingInstrumentor._instrumented_loggers = False
+
+            del logging.Logger.propagate
+
+            for uninstrument_logger in chain(
+                (logging.getLogger(),),
+                (
+                    logging.getLogger(name)
+                    for name in logging.root.manager.loggerDict
+                ),
+            ):
+                if not isinstance(uninstrument_logger, logging.Logger):
+                    continue
+
+                if hasattr(uninstrument_logger, "_otel_propagate"):
+                    uninstrument_logger.propagate = (
+                        uninstrument_logger._otel_propagate
+                    )
+                    del uninstrument_logger._otel_propagate
+                if hasattr(uninstrument_logger, "_otel_handler"):
+                    if uninstrument_logger._otel_handler is not None:
+                        uninstrument_logger.removeHandler(
+                            uninstrument_logger._otel_handler
+                        )
+                    del uninstrument_logger._otel_handler